SPIDeviceWithRegisters has been cut. But IMUs are broken.

+++ src/Digitabulum/LSM9DS1/IIU.cpp
@@ -338,7 +338,7 @@ void IIU::assign_legend_pointers(void* a,
 
 
 void IIU::assign_register_pointers(IMURegisterPointers* _reg) {
-  for (int i = 0; i < sizeof(IMURegisterPointers); i+=4) {
+  for (size_t i = 0; i < sizeof(IMURegisterPointers); i+=4) {
     *(&_reg_ptrs + i) = *(_reg + i);
   }
 }
